Made the Gui chatbot functional.
I removed some useless code. I added a popup message for when the program starts. If "quit" is typed the program will end. Refactored the passTestButton to the checkerButton and added string to the constructor call of it. Changed the layout of the GUI a bit. Made it so that the enter button also submitted the user's input.